Fix bug where @ type="xhtml" was shown in News
Most feed entries had a <summary> tag with no attributes, but SRFI
feed entries had <summary type="xhtml">. This is represented in SXML
as (summary (@ (type "xhtml")) ...), and taking the cdr of that caused
the `@ (type "xhtml")` to be visible on the page.

Fix by skipping any (@ ...) attributes in the SXML summary element.