Retire the (people ...) section
It has never been used for anything.

The original intent was to map people's real names to Unix usernames.
But the set of names ended up being arbitrary. There is no logic as to
who is listed and who is not. And the listed usernames do not reliably
correspond to actual usernames on any given system.