Retire wwwlocal script
This was originally used to start a localhost-only web server to test
changes to www.scheme.org without pushing them to any remote server.
But now that we have two domains and subdomains sourced from this
repo, things are complicated and the script has stopped working.

Since we can easily make staging subdomains to test changes before
pushing them to production, the script is not terribly useful. And
making it work locally with multiple domains would probably involve
editing /etc/hosts. It's not worth the effort. Let's drop it.