Move most sites to tuonela
With this commit, the server name alpha.servers.scheme.org is retired.
This server is now just scheme.org. In our DNS records, we refer to it
using the standard zone file syntax "@" which refers to the origin of
the domain.

Since there are no servers remaining under *.servers.scheme.org, the
whole subdomain servers.scheme.org is retired too.

To recap, our servers are now:

- scheme.org (the origin server)
- ironwolf.scheme.org
- tuonela.scheme.org

With this commit, all static sites are moved to tuonela except for the
following:

- www
- www.staging
- redirect
- lists
- files

These stay on the origin server (nee alpha).

lists stays because it involves mailing lists, which we have a
persistent TODO item to explore how to do.

files stays because tuonela doesn't (yet) have enough disk space to
house the massive backlog of Scheme releases. It should move as soon
as we acquire extra disk space.