Convert doc/ from Markdown to HTML
Long overdue. We no longer need a Markdown parser, which tends to be a
tricky dependency. We'd need an XML parser for the Atom feed anyway.

The HTML fragments in the doc directory start with <?xml?> since the
Chicken XML parser needs that. The parser is also confused by
<!DOCTYPE> (which html-tidy likes to insert unconditionally). We
should sort this out later, but the current situation is livable.