Add source information for projects
It is hard for people to figure out where the source code for each
piece of Scheme.org is found. The site is big enough that we need more
than one kind of navigation aid. Start by recording the version
control repository for each project.

When a project is stored in a subdirectory of a repository, give the
subdirectory in the "path" property.

The source information is meant to be displayed to humans. A computer
program will not be able to reconstruct a project on this basis alone.